requirements.txt
python-telegram-bot==13.7
openai==0.27.0
python-dotenv==0.19.1
